POC: client-side RC4 encryption.
I added an `--enc-key-file KEY_FILE` flag, where KEY_FILE stores
the RC4 key (key must be >= 1 byte and <= 256 bytes).  I
modified the ReadFile and WriteFile FUSE callbacks to invoke XOR
the data with the RC4 stream.  These callbacks take care of
XOR'ing with the correct offset in the stream.