POC for AES CTR clien-side-encryption.
Use the `--enc-key-file KEY_FILE` and `--iv-file IV_FILE` command-line
options to `gcsfuse` to specify the file's that have the encryption key
and IV.  The encryption key can be 16, 24, or 32 bytes-long; the IV must
be 16-bytes.  Both options must be specified together; if neither
option is specified, `gcsfuse` operates normally.